requirements.txt
python-telegram-bot==20.7
web3==6.11.1
sqlalchemy==2.0.23
cryptography==41.0.5
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
eth-account==0.10.0
eth-typing==3.5.0
eth-utils==2.3.0 